+++ crates/swc_config/src/config_types/bool_or_object.rs
