+++ crates/swc_config/src/config_types/mod.rs
@@ -1,4 +1,4 @@
-pub use self::{bool_config::BoolConfig, bool_or::BoolOr};
+pub use self::{bool_config::BoolConfig, bool_or_object::BoolOr};
 
 mod bool_config;
-mod bool_or;
+mod bool_or_object;
